requirements.txt
mss
colorama
keyboard
pillow
